move `useSidebarSearch.ts` to its own sub-directory for better separation of concerns
