+++ src/App/components/Sidebar/SidebarSearchResults.tsx
@@ -1,12 +1,12 @@
 import styles from './SidebarSearchResults.module.css';
-import PoolsSearchResults from './PoolsSearchResults/PoolsSearchResults';
-import PositionsSearchResults from './PositionsSearchResults/PositionsSearchResults';
-import OrdersSearchResults from './OrdersSearchResults/OrdersSearchResults';
-import TxSearchResults from './TxSearchResults/TxSearchResults';
-import { PoolStatsFn } from '../../../functions/getPoolStats';
-import { sidebarSearchIF } from '../useSidebarSearch/useSidebarSearch';
-import { useAppSelector } from '../../../../utils/hooks/reduxToolkit';
-import { TokenPriceFn } from '../../../functions/fetchTokenPrice';
+import PoolsSearchResults from './SidebarSearchResults/PoolsSearchResults/PoolsSearchResults';
+import PositionsSearchResults from './SidebarSearchResults/PositionsSearchResults/PositionsSearchResults';
+import OrdersSearchResults from './SidebarSearchResults/OrdersSearchResults/OrdersSearchResults';
+import TxSearchResults from './SidebarSearchResults/TxSearchResults/TxSearchResults';
+import { PoolStatsFn } from '../../functions/getPoolStats';
+import { sidebarSearchIF } from './useSidebarSearch/useSidebarSearch';
+import { useAppSelector } from '../../../utils/hooks/reduxToolkit';
+import { TokenPriceFn } from '../../functions/fetchTokenPrice';
 
 interface propsIF {
     cachedPoolStatsFetch: PoolStatsFn;
